Update prefs.js add 5 minutes value
I need to use 5 minutes because 30 minutes is too long. I have a wireless keyboard and mouse and I block the screen and put them away from the table. If you do not use the extension, then the screen wakes up from an accidental press, and half an hour is too much. One minute would have been enough for me, but this is already too ugly for me, so please make 5 minutes - it will just be like turning off the instant screen off. I do this at home, but the auto-update of the extension spoils it. Thanks a lot!